Fix keyboard ISR placing invalid value in buffer
When running the mouse test in Steem, if the mouse was shaken too much and too fast, then the keyboard would no longer be responsible. This was caused by a lack of checking of what scancodes were being passed and an assumption that the scancode given would never be zero.

Note that the game itself still has issues with the keyboard and mouse being unresponsive when running under Steem.