Update bitmaps, events, font16, model, and renderer modules
The model needed to be updated to take advantage of the functionality
added by the move module. As a result of these changes, the other modules
had to be updated as well. Since they had to be updated anyway, various
fixes and improvements were made. Extraneous includes were removed from
the model and the bitmaps and font16 module were heavily modified to
promote abstraction.