Improve player movement among other things (#29)
Adds a queue so that all player movement is kept instead of being discarded if the player is set to move. Also, to make the player easier to control, if the user requests that the player moves in the opposite direction from where the player is set to move, the movement queue is cleared. This allows the user to easily cancel any movements.

The model needed to be updated to take advantage of the functionality added by the move module. As a result of these changes, the other modules had to be updated as well. Since they had to be updated anyway, various fixes and improvements were made. Extraneous includes were removed from the model and the bitmaps and font16 module were heavily modified to promote abstraction.

Additionally, a bus error was fixed in tst_mdl.c.