Replace all usage of osbind.h (#36)
One of the most significant changes made has been to create a super module to replace usage of osbind in regards to using Super(). This fixes #33. Also, all public-facing functions no longer require super privileges. Using the new functionality from the super module, functions that require super can tell if they have the required privileges and grant itself them if necessary.

Additionally, the testing functionality has been completely overhauled to be more portable and not depend on osbind. A lots of the tests have been revamped as a result. What used to be tst_rast has been split up into tst_shp and tst_bmp, as the original file became too big to be compiled on TOS. The bitmap arrays that used to exist in tst_bmp was moved out into the bitmaps module in order to encourage future use of the images.

The ArgList module was added to provide an easy way to pass arguments to test functions.

All assembly files now end with '_asm' to allow for future expansion in case some C code should be added to a module.

While all the other changes were being made, renderGame() in src/acaddom.c was simplified. It kept on masking and unmasking interrupts presumably as a failed attempt at optimization.

Furthermore, a bunch of functions that used to exist in test modules were moved out so they could be used by other modules. Fixes #30.

Finally, issue #31 was tackled quite a bit. The function getBlockingKey() was removed from a test module and equivalent functions added to the input module.