.txt -> .adoc
Some AsciiDoc documents were created with a .txt extension so they
could at least be readable in plain text on TOS which only supports a
maximum of three characters for a file extension. Since native
compilation support for TOS will be removed in the future, this is less
of a concern. Having programs recognize the files as AsciiDoc will
improve readability of the documentation.