Refactor renderer (#59)
The file documentation for the renderer module has been updated to better reflect what the module does. Also, constants have been added to make it easier to reference the height and width of the font currently used for the label. Furthermore, the Lives Box and Score Box render their own labels. As a result of the change, the screen buffer is copied instead of having renderPlayArea() called for each buffer.