Fix bugs and improve quality in input functions (#65)
A mouse regression has been fixed that caused the position of a cursor to be treated as new if the mouse was left or right clicked with. Fixes #64.

Also, the process of restoring the original keyboard ISR has been improved. The original ISR will only be restored when no keys are pressed.