Refactor menu code (#49)
A few features have been added. For instance, a function has been added to check if the cursor is currently rendering. Also, the cursor is hidden on the Title and Game Over Screens when using the keyboard to cycle between buttons.

However, the most major change was the removal of duplicate code from the Game Over and Title Screen by creating a Menu object. The Menu object also serves to abstract code away. Most Button and InfoBar functions were modified to require the usage of a Menu object. Many new Button and Score functions were created as a result of the addition of the Menu and related refactoring.

The details about the Title Screen and Game Over Screen were combined in the MANUAL in order to reduce the amount of duplicate information and also to reflect the fact that both menus share a lot of code between each other.

During the process of fixing bugs, the logic for quitting to the Title Screen was simplified. Also, the maximum length of a Label was changed to 35. That was so the resultant LabelStr would contain an even number of characters. It was done as part of an attempt at fixing an issue with the compiler using the wrong offsets on (and possibly corrupting) the stack.