Fix errant code (#68)
Contains several bug fixes and code quality improvements.

cmpScore() was changed to use SINT32 as its return type instead of UINT32, as it can return a negative integer.

Also, initInfoBar() and addInfoBar() have been combined into one function. As part of that process, extra checks were added and a bug was fixed that would have prevented the maximum allowable number of strings to be added to an InfoBar. A similar combination of functions occurred with initButton() and addButton().

Additionally, instead of initHazard() initializing parts of rows that are unused, they are now initialized when needed by addHazard(). This ensures that the proper orientation is set for hazards added to rows after the latter has been initialized.

Finally, a global Menu object has been created, as alignment issues occur when the object is placed onto the stack.